Refactoring Controllers, Factory, Services
