+++ app/Http/Controllers/Api/TaskIndexController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\TaskIndexRequest;
+use App\Services\AnswerService;
 use App\Services\TaskIndexService;
 use Database\Factories\TaskDataFactory;
 use Exception;
@@ -11,16 +12,17 @@ use Illuminate\Http\JsonResponse;
 
 class TaskIndexController extends Controller
 {
-    use ControllerTrait, AnswerTrait;
+    use ControllerTrait;
 
     /**
      * @param TaskIndexService $taskIndexService
      * @param TaskDataFactory $taskFactory
+     * @param AnswerService $answerService
      */
     public function __construct(
         protected TaskIndexService $taskIndexService,
         protected TaskDataFactory  $taskFactory,
-
+        protected AnswerService    $answerService,
     )
     {
     }
@@ -36,7 +38,7 @@ class TaskIndexController extends Controller
         try {
             $response = $this->taskIndexService->index($validData);
 
-            $this->setAnswer($response);
+            $this->answerService->setAnswer($response);
 
         } catch (Exception $e) {
             $this->getCatch($e);
