+++ database/factories/AnswerDataFactory.php
@@ -7,24 +7,11 @@ use App\Data\AnswerData;
 class AnswerDataFactory
 {
     /**
-     * @param int $status
-     * @param string $message
-     * @param bool|object|array|null $data
-     * @param int|string|null $code
+     * @param array $data
      * @return AnswerData
      */
-    public static function answerData(
-        int                    $status,
-        string                 $message,
-        bool|object|array|null $data,
-        int|string|null        $code
-    ): AnswerData
+    public static function answerData(array $data): AnswerData
     {
-        return new AnswerData(
-            status: $status,
-            message: $message,
-            data: $data,
-            code: $code
-        );
+        return new AnswerData(...$data);
     }
 }
